+++ packages/domains/src/indices/tx.ts
@@ -1,26 +1,28 @@
-<!--
-// Copyright © 2020 Anticrm Platform Contributors.
-// 
+//
+// Copyright © 2020, 2021 Anticrm Platform Contributors.
+//
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
 // obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// 
+//
 // See the License for the specific language governing permissions and
 // limitations under the License.
--->
-<script lang="ts">
-  import { CreateTx, Space, CORE_CLASS_CREATE_TX } from '@anticrm/domains'
-  import { Tx } from '@anticrm/core'
+//
 
-  export let tx: Tx
+import { DomainIndex, Storage, TxContext, Tx } from '@anticrm/core'
 
-  function spaceName() {
-    return (((tx as CreateTx).object as unknown) as Space).name
+export class TxIndex implements DomainIndex {
+  private storage: Storage
+
+  constructor (storage: Storage) {
+    this.storage = storage
   }
-</script>
 
-{#if tx._class === CORE_CLASS_CREATE_TX}Создал пространство <b>{spaceName()}</b>{/if}
+  tx (ctx: TxContext, tx: Tx): Promise<any> {
+    return this.storage.store(ctx, tx)
+  }
+}
