+++ plugins/chunter/src/components/internal/ActivityView.svelte
@@ -13,10 +13,10 @@
 // limitations under the License.
 -->
 <script type='ts'>
-  import { Ref, Tx, CORE_CLASS_CREATE_TX } from '@anticrm/core'
+  import { Ref, Tx } from '@anticrm/core'
   import { onDestroy } from 'svelte'
   import { WorkbenchApplication } from '@anticrm/workbench'
-  import { Space } from '@anticrm/core'
+  import { Space, CORE_CLASS_CREATE_TX } from '@anticrm/domains'
 
   import ScrollView from '@anticrm/sparkling-controls/src/ScrollView.svelte'
   import ActivityItem from './ActivityItem.svelte'
