+++ plugins/chunter/src/components/internal/DateItem.svelte
@@ -11,6 +11,7 @@
   //
   // See the License for the specific language governing permissions and
   // limitations under the License.
+
   import chunter from '@anticrm/chunter'
   import Icon from '@anticrm/platform-ui/src/components/Icon.svelte'
