+++ plugins/task/src/__model__.ts
@@ -21,7 +21,8 @@ import { UX } from '@anticrm/presentation/src/__model__'
 import presentation from '@anticrm/presentation'
 import workbench from '@anticrm/workbench/src/__model__'
 import chunter from '@anticrm/chunter/src/__model__'
-import { Application, Ref, Property, StringProperty } from '@anticrm/core'
+import { Ref, Property, StringProperty } from '@anticrm/core'
+import { Application } from '@anticrm/domains'
 import { TVDoc } from '@anticrm/model/src/__model__'
 
 export const DOMAIN_TASK = 'task'
