+++ plugins/workbench/src/components/internal/spaces/utils.ts
@@ -1,4 +1,20 @@
-import { Space, SpaceUser, Property, StringProperty } from '@anticrm/core'
+//
+// Copyright © 2020, 2021 Anticrm Platform Contributors.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import { Property, StringProperty } from '@anticrm/core'
+import { Space, SpaceUser } from '@anticrm/domains'
 import { CoreService } from '@anticrm/platform-core'
 
 /**
