+++ plugins/workbench/src/index.ts
@@ -18,7 +18,8 @@ import { plugin, Plugin, Service } from '@anticrm/platform'
 import core from '@anticrm/platform-core'
 import ui, { AnyComponent, Asset } from '@anticrm/platform-ui'
 import { IntlString } from '@anticrm/platform-i18n'
-import { Application, VDoc, Doc, Class, Ref } from '@anticrm/core'
+import { Ref, Class, Doc } from '@anticrm/core'
+import { Application, VDoc } from '@anticrm/domains'
 
 export interface Perspective extends Doc {
   label: IntlString
